627228ing

+++ code/其他/test/test.cpp
@@ -61,7 +61,7 @@ signed main() {
     auto qu=[&](int l,int r) {
         if(l>r) return 0;
         int lg=__lg(r-l+1);
-        return __gcd(s[lg][l],s[lg][r-(1<<lg)+1]);
+        return (int)__gcd(s[lg][l],s[lg][r-(1<<lg)+1]);
     };
     int an=1;
     REP1(x2,n-1) {
